pyproject.toml
[tool.poetry]
name = "test-task-notification"
version = "0.1.0"
description = ""
authors = ["Julia Bengert <yubengert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.1"
django = "^5.2.4"
djangorestframework = "^3.16.0"
celery = "^5.5.3"
django-celery-beat = "^2.8.1"
redis = "^6.3.0"
gevent = "^25.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
